+++ test/plugins/test_query_history.py
@@ -79,7 +79,7 @@ class TestQueryHist:
             from opentelemetry.sdk.resources import Resource
 
             def _get_spans(self, resource: Resource) -> Spans:
-                return TestSpans(resource, self._configuration)
+                return TestSpans(resource, self._configuration, self._otel_manager)
 
         def __local_get_plugin_class(source: str):
             return TestQueryHistoryPlugin
@@ -161,7 +161,7 @@ class TestQueryHist:
             from opentelemetry.sdk.resources import Resource
 
             def _get_spans(self, resource: Resource) -> Spans:
-                return TestSpans(resource, self._configuration)
+                return TestSpans(resource, self._configuration, self._otel_manager)
 
         def __local_get_plugin_class(source: str):
             return TestQueryHistoryPlugin
@@ -232,7 +232,7 @@ class TestQueryHist:
             from opentelemetry.sdk.resources import Resource
 
             def _get_spans(self, resource: Resource) -> Spans:
-                return TestSpans(resource, self._configuration)
+                return TestSpans(resource, self._configuration, self._otel_manager)
 
         def __local_get_plugin_class(source: str):
             return TestQueryHistoryPlugin
@@ -317,7 +317,7 @@ class TestQueryHistDdl:
             from opentelemetry.sdk.resources import Resource
 
             def _get_spans(self, resource: Resource) -> Spans:
-                return TestSpans(resource, self._configuration)
+                return TestSpans(resource, self._configuration, self._otel_manager)
 
         def __local_get_plugin_class(source: str):
             return TestQueryHistoryPlugin