+++ test/plugins/test_query_history_no_admin.py
@@ -90,7 +90,7 @@ class TestQueryHistNoAdmin:
             from opentelemetry.sdk.resources import Resource
 
             def _get_spans(self, resource: Resource) -> Spans:
-                return TestSpans(resource, self._configuration)
+                return TestSpans(resource, self._configuration, self._otel_manager)
 
         def __local_get_plugin_class(source: str):
             return TestQueryHistoryNoAdminPlugin